scripts/requirements.txt
loguru
shapely
protobuf==3.20.2
omegaconf==2.3.0
hydra-core==1.3.2
rl-games==1.6.1
expiringdict==1.2.2
gymnasium
tqdm
joblib
bottle
pandas
fire
seaborn
textual
pipetools
tinydb
reloading
pysnooper